Add files via upload
This project works as a ocr model which detects number plates and notifies the owner about the vehicles whereabouts and other stuffs. 
Python stack is used for building both the ends .
the security measure are as the both caller and receiver are connected with a temporary number by the back end.